refactor(cleanup): fold URLs into core/urls.py, drop dead indirection [AE-2946]
Self-review cleanup:

- Move CONSOLE_BASE_URL / CONSOLE_URL from core/resources/constants.py
  into core/urls.py — they are Runpod service URLs like the others, with
  the same env-override pattern, and splitting them across two files had
  no principled basis.
- Drop the dead __all__ and cargo-cult 'backward compat' comment from
  core/api/runpod.py. __all__ only affects wildcard imports; it did
  nothing for the actual test imports.
- Remove the unused RUNPOD_API_BASE_URL import from core/api/runpod.py.
  It was only imported to populate that fake __all__.
- Point tests/unit/core/api/test_runpod_graphql_extended.py at
  core/urls.py for RUNPOD_REST_API_URL instead of re-exporting through
  core/api/runpod.py.
- Collapse the 3-line os.environ.get(...).rstrip() split in
  stubs/dependency_resolver.py codegen into a single readable line.
  rstrip dropped — the default is clean and the env is internal.
