+++ src/runpod_flash/core/api/runpod.py
@@ -12,12 +12,9 @@ import aiohttp
 
 from runpod_flash.core.credentials import get_api_key
 from runpod_flash.core.exceptions import RunpodAPIKeyError
-from runpod_flash.core.urls import GRAPHQL_URL, RUNPOD_API_BASE_URL, RUNPOD_REST_API_URL
+from runpod_flash.core.urls import GRAPHQL_URL, RUNPOD_REST_API_URL
 from runpod_flash.core.utils.backoff import BackoffStrategy, get_backoff_delay
 from runpod_flash.runtime.exceptions import GraphQLMutationError, GraphQLQueryError
-
-# Re-exported for backward compat with tests that import these from here.
-__all__ = ["RUNPOD_API_BASE_URL", "RUNPOD_REST_API_URL", "RunpodGraphQLClient"]
 
 log = logging.getLogger(__name__)
