+++ src/runpod_flash/core/resources/serverless.py
@@ -24,8 +24,8 @@ from ..exceptions import RunpodAPIKeyError
 from ..utils.backoff import get_backoff_delay
 from .base import DeployableResource
 from .cloud import runpod
+from ..urls import CONSOLE_URL
 from .constants import (
-    CONSOLE_URL,
     DEFAULT_WORKERS_MAX,
     DEFAULT_WORKERS_MIN,
     validate_python_version as _validate_python_version,
